my_ui/pyside6_ui/requirements.txt
PySide6>=6.4.0
PyPDF2>=3.0.0
faiss-cpu>=1.7.0
numpy>=1.20.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0